[Fix] - Views and querys

+++ Model/TopCompanyResult.cs
@@ -10,6 +10,6 @@ namespace PagoAgilFrba.Model
         public String name { get; set; }
         public String address { get; set; }
         public String cuit { get; set; }
-        public int value { get; set; }
+        public Decimal value { get; set; }
     }
 }
